Added out of bounds scale
