+++ src/components/Player/PlayerMenu.tsx
@@ -24,7 +24,7 @@ import { useTheme } from '$lib/Hooks/useTheme';
 import { SearchedSongType } from '$types/GeneralTypes';
 import { FileElement, FilePicker } from '$cmp/Inputs/FilePicker';
 import "./menu.css"
-import { ThemeStore } from '$stores/ThemeStore/ThemeProvider';
+import { Theme } from '$stores/ThemeStore/ThemeProvider';
 import { KeyboardEventData, KeyboardProvider } from '$lib/Providers/KeyboardProvider';
 import { hasTooltip, Tooltip } from "$cmp/Utility/Tooltip"
 import { HelpTooltip } from '$cmp/Utility/HelpTooltip';
@@ -399,7 +399,7 @@ function Menu({ functions, data }: MenuProps) {
 
 interface SongRowProps {
     data: SongStorable
-    theme: ThemeStore
+    theme: Theme
     folders: Folder[]
     functions: {
         removeSong: (name: string, id: string) => void
